test fluorescence and refactor compute_tf tests
